Move test dock regression to separated file.
